// src/app/commons/enums/exercise-category-groups.enum.ts
export enum ExerciseCategoryGroups {

  CHEST = 'CHEST',
  SHOULDERS = 'SHOULDERS',
  BICEPS = 'BICEPS',
  TRICEPS = 'TRICEPS',
  FOREARMS = 'FOREARMS',
  UPPER_BACK = 'UPPER_BACK',
  LOWER_BACK = 'LOWER_BACK',
  ABS = 'ABS',
  GLUTES = 'GLUTES',
  THIGH_FRONT = 'THIGH_FRONT',
  THIGH_BACK = 'THIGH_BACK',
  CALVES = 'CALVES',
  CARDIO = 'CARDIO',
  CUSTOM = 'CUSTOM',
  FAVOURITE = 'FAVOURITE'
}